+++ CHANGELOG.md
@@ -1,7 +1,10 @@
-# 0.0.6 (????-??-??)
+# 1.0.0 (????-??-??)
 
 * Added `ports` keyword argument to `launch` method for `ROSWire`, allowing
   users to specify an optional container-host port mapping.
+* Added required `sources` positional argument to `launch` method for `ROSWire`
+  and most methods for `DescriptionManager` to allow user to specify the setup
+  files that should be used by the application.
 
 
 # 0.0.5 (2019-08-28)